fix(root): update prop description backticks to be code instead
Update prop description backticks to be code instead as intended

. #263
