--- src/utils/helpers.ts
@@ -1,10 +1,0 @@
-// eslint-disable-next-line import/prefer-default-export
-export function debounce(fn: Function, timeout = 300) {
-  let timer: ReturnType<typeof setTimeout>;
-  return function func(this: any, ...args: any[]) {
-    clearTimeout(timer);
-    timer = setTimeout(() => {
-      fn.apply(this, args);
-    }, timeout);
-  };
-}
